support for saved games
works by chroot'ing to a private directory for each game.  send the
command "save" (game will prompt for a filename), then send a filename
as if it were a command.  Similarly, to restore, send "restore", then
send the filename.  The filename should not have slashes, but if a bad
guy tries that, the chroot will prevent problems.